+++ packages/sdk-core/src/ERC20Token.ts
@@ -1,7 +1,7 @@
 import {
     IERC20Metadata,
     IERC20Metadata__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+} from "./typechain-types";
 import { ethers } from "ethers";
 
 import Operation from "./Operation";