+++ packages/sdk-core/src/Framework.ts
@@ -4,7 +4,7 @@ import {
     Resolver__factory,
     Superfluid__factory,
     SuperfluidLoader__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+} from "./typechain-types";
 import { ethers } from "ethers";
 import Web3 from "web3";