+++ packages/sdk-core/src/Governance.ts
@@ -1,7 +1,7 @@
 import {
     SuperfluidGovernanceII,
     SuperfluidGovernanceII__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+} from "./typechain-types";
 import { ethers } from "ethers";
 
 import {