+++ packages/sdk-core/src/Host.ts
@@ -1,7 +1,7 @@
 import {
     Superfluid,
     Superfluid__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+} from "./typechain-types";
 import { ethers, Overrides } from "ethers";
 
 import Operation from "./Operation";