+++ packages/sdk-core/src/InstantDistributionAgreementV1.ts
@@ -1,7 +1,7 @@
 import {
     IInstantDistributionAgreementV1,
     IInstantDistributionAgreementV1__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+} from "./typechain-types";
 import { ethers } from "ethers";
 
 import Host from "./Host";