+++ packages/sdk-core/src/Operation.ts
@@ -1,4 +1,4 @@
-import { Superfluid__factory } from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
+import { Superfluid__factory } from "./typechain-types";
 import { ethers } from "ethers";
 
 import {